+++ docs/demo/demo-custom.js
@@ -27,7 +27,7 @@ new mar10.Wunderbaum({
     },
     { id: "description", title: "Details", width: "*" },
   ],
-  resizableColumns: true,
+  columnsResizable: true,
   types: {
     category: { colspan: true },
     electronics: { icon: "bi bi-cpu" },