+++ test/playground.js
@@ -73,7 +73,7 @@ const tree = new Wunderbaum({
   // source: "generator/ajax_1k_3_54 t_c.json",
   // source: "generator/fixture_department_1k_3_6_flat_comp.json",
   // source: "generator/fixture_department_1k_3_6_comp.json",
-  resizableColumns: true,
+  columnsResizable: true,
 
   // source: "../docs/assets/ajax-tree-products.json",
   // source: